backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic[email]==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyNaCl==1.5.0
scikit-learn==1.3.2
joblib==1.3.2
numpy==1.26.2
pandas==2.1.4
python-multipart==0.0.6
bcrypt==4.1.2
